chore: restruct directory for more friendly readability
