+++ service/poly2zil_test.go
@@ -9,6 +9,7 @@ import (
 )
 
 var p *PolySyncManager
+var s *ZilSender
 
 func init() {
 	zilSdk := provider.NewProvider("https://polynetworkcc3dcb2-5-api.dev.z7a.xyz")
@@ -18,17 +19,24 @@ func init() {
 			CrossChainManagerContract: "zil1ur4vwcmcz3jqypksgq7qeju2sk5jrskzaadau5",
 		}},
 	}
+	s = &ZilSender{
+		cfg: &config.Config{ZilConfig: &config.ZILConfig{
+			CrossChainManagerContract: "zil1ur4vwcmcz3jqypksgq7qeju2sk5jrskzaadau5",
+		}},
+		zilSdk: zilSdk,
+		acc:    "zil1ur4vwcmcz3jqypksgq7qeju2sk5jrskzaadau5",
+	}
 }
 
 func TestPolySyncManager_FindLatestHeight(t *testing.T) {
 	fmt.Println(p.findLatestHeight())
 }
 
-func TestPolySyncManager_CheckIfFromChainTxExist(t *testing.T) {
-	exist := p.checkIfFromChainTxExist(3, "0x00ca93f8738111a063d8ab7221f47c70a4cade0ca4a2829df494cd4b5e231bd6")
+func TestZilSender_CheckIfFromChainTxExist(t *testing.T) {
+	exist := s.checkIfFromChainTxExist(3, "0x00ca93f8738111a063d8ab7221f47c70a4cade0ca4a2829df494cd4b5e231bd6")
 	assert.Equal(t, exist, true)
 
-	exist = p.checkIfFromChainTxExist(3, "0x00ca93f8738111a063d8ab7221f47c70a4cade0ca4a2829df494cd4b5e231bd7")
+	exist = s.checkIfFromChainTxExist(3, "0x00ca93f8738111a063d8ab7221f47c70a4cade0ca4a2829df494cd4b5e231bd7")
 	assert.Equal(t, exist, false)
 
 }
